refactor: 127-refactor code
