+++ src/components/profile/layout.ts
@@ -3,7 +3,7 @@ import { createDiv, createH2 } from '../../utils/create-elements/create-tags';
 import { PROFILE_CLASSES } from './constants';
 import { createPasswordSection } from './password-section/password-section';
 import { createPersonalInfoSection } from './personal-info-section/personal-info';
-import { createOptionalAddressSection } from './optional-shipping-section/create-optional-address-section';
+import { createOptionalAddressSection } from './optional-address-section/create-optional-address-section';
 import { createDefaultAddressSection } from './default-address-section/create-default-address-section';
 
 export async function profileLayout(): Promise<HTMLDivElement | void> {
