+++ src/helpers/buttons-emitter.ts
@@ -41,7 +41,6 @@ function clearInputValues(inputs: HTMLInputElement[]): void {
   }
 }
 
-// eslint-disable-next-line max-lines-per-function
 export function activateButtonEmitter(
   emitter: CustomEventEmitter,
   buttons: HTMLButtonElement[],
@@ -71,7 +70,6 @@ export function activateButtonEmitter(
     if (emitter === personalInfoEmitter) {
       await resetInputDisplayFromServer(inputs);
     }
-
     if (emitter === passwordEmitter) {
       clearInputValues(inputs);
     }
@@ -79,15 +77,12 @@ export function activateButtonEmitter(
     if (emitter === defaultShippingAddressEmitter && select instanceof HTMLSelectElement) {
       await resetDefaultAddressInputFromServer(inputs, select, 'shipping');
     }
-
     if (emitter === defaultBillingAddressEmitter && select instanceof HTMLSelectElement) {
       await resetDefaultAddressInputFromServer(inputs, select, 'billing');
     }
-
     if (emitter === firstOptionalAddressEmitter && select instanceof HTMLSelectElement) {
       await resetOptionalAddressInputFromServer(inputs, select, 'optional-shipping');
     }
-
     if (emitter === secondOptionalAddressEmitter && select instanceof HTMLSelectElement) {
       await resetOptionalAddressInputFromServer(inputs, select, 'optional-billing');
     }
